Cat: Do not send TR immediately for Display_Text
Currently for DISPLAY_TEXT proactive command, if the application
response is not required, CatService sends a terminal response to
card immediately and then sends the proactive command to StkApp.
On receiving the TR, the card responds with a Session_End. If the
Session_End is sent to StkApp before it finishes processing the
Display_Text command, a nullPointerException is seen in StkApp

Do not send Display_Text TR to card immediately. Send the TR after
the Display_Text command response has been received from StkApp.

Change-Id: Ic00067fe48b264c77092dfef3278472d95663016
CRs-Fixed: 608287